==> notebooks/02_DEP_02-Preprocess_user_data.ipynb <==
# Import thư viện cần thiết

import pandas as pd
import numpy as np
import os
import sys

# Khám phá dữ liệu về video

## Đọc dữ liệu video sau khi đã được xử lý

video_df = pd.read_parquet("preprocessed_videos.parquet")
video_df.info()

## Danh sách các user được đề cập trong dữ liệu về video

unique_author_ids = video_df["author.id"].unique()
print(f"Number of unique author ids: {len(unique_author_ids)}")

# Tiền xử lý dữ liệu về user

## Đọc dữ liệu user

Xác định trước kiểu dữ liệu phù hợp cho các cột trong dữ liệu về người dùng

user_dtypes = {
    "user.bioLink.risk": np.object_,
    "user.commerceUserInfo.category": np.object_,
    "user.commerceUserInfo.categoryButton": np.object_,
    "user.downloadSetting": np.object_,
    "user.duetSetting": np.object_,
    "user.followingVisibility": np.object_,
    "user.id": np.object_,
    "user.nickNameModifyTime": np.object_,
    "user.nickname": np.object_,
    "user.profileEmbedPermission": np.object_,
    "user.profileTab.showQuestionTab": np.object_,
    "user.relation": np.object_,
    "user.secUid": np.object_,
    "user.signature": np.object_,
    "user.stitchSetting": np.object_,
    "user.uniqueId": np.object_,
    "user.bioLink.link": np.object_,
    "user.roomId": np.object_,
}

user_df = pd.read_csv("data/interim/final_raw_users.csv", dtype=user_dtypes)
user_df.info()

## Chỉ giữ lại các user có trong dữ liệu về video

# Check if the user ids in the video_df are present in the user_df
missing_user_ids = set(video_df["author.id"]) - set(user_df["user.id"])
if missing_user_ids:
    print(f"Missing user ids in user_df: {missing_user_ids}")
else:
    print("All user ids in video_df are present in user_df.")

## Loại bỏ các cột bị thiếu quá nhiều dữ liệu

Tính tỷ lệ thiếu giá trị của các cột

# Calculate the missing rate for each column in user_df
missing_rate = user_df.isnull().mean()
missing_rate = missing_rate[missing_rate > 0].sort_values(
    ascending=False).reset_index()
missing_rate.columns = ["column_name", "missing_rate"]
missing_rate["missing_rate"] = missing_rate["missing_rate"].apply(
    lambda x: f"{x:.2%}")
missing_rate = missing_rate.set_index("column_name")
# Print the missing rate for each column in user_df
print("Missing rate for each column in user_df:")
missing_rate

Loại bỏ tất cả các cột bị thiếu dữ liệu

# Remove all columns with more than 50% missing values
user_df = user_df.dropna(thresh=len(user_df) * 0.5, axis=1)

# Calculate the missing rate for each column in user_df
missing_rate = user_df.isnull().mean()
missing_rate = missing_rate[missing_rate > 0].sort_values(
    ascending=False).reset_index()
missing_rate.columns = ["column_name", "missing_rate"]
missing_rate["missing_rate"] = missing_rate["missing_rate"].apply(
    lambda x: f"{x:.2%}")
missing_rate = missing_rate.set_index("column_name")
# Print the missing rate for each column in user_df
print("Missing rate for each column in user_df:")
missing_rate

Loại bỏ các cột chỉ có duy nhất một giá trị

# Loại bỏ các cột chỉ có một giá trị duy nhất
user_df = user_df.loc[:, user_df.nunique() > 1]
user_df.info()

Điền giá trị "others" cho các cột có giá trị là "unknown"

# Điền giá trị "others" cho các cột không có kiểu dữ kiệu số 
non_numeric_columns = user_df.select_dtypes(exclude=[np.number]).columns
for col in non_numeric_columns:
    user_df[col] = user_df[col].fillna("others")

## Kết quả sau khi tiền xử lý dữ liệu về user

user_df.info()

## Feature Engineering cho dữ liệu về user

### Tính số lượng video trung bình mà 1 user đăng tải trong mỗi tuần

# Group DataFrame by author.id and count the number of videos for each author
author_video_counts = video_df.groupby(
    "author.id").size().reset_index(name="video_count")

# Get the time of the first and last video for each author
author_first_last = video_df.groupby("author.id")[
    "createTime"].agg(["min", "max"]).reset_index()

# Merge the two DataFrames to get the video count along with the first and last video times
author_video_info = pd.merge(
    author_video_counts, author_first_last, on="author.id")

# Calculate the number of weeks between the first and last video
author_video_info["weeks_between"] = (
    author_video_info["max"] - author_video_info["min"]).dt.days / 7

# Calculate the average number of videos per week for each author
author_video_info["avg_videos_per_week"] = (
    author_video_info["video_count"] / author_video_info["weeks_between"]).fillna(0)

author_video_info.head()

author_video_info["avg_videos_per_week"].describe()

Chỉ giữ lại các cột cần thiết và đổi tên cột cho phù hợp

# Rename the columns to match the user_df
author_video_info = author_video_info.rename(
    columns={"author.id": "user.id"})

# Keep only the relevant columns
author_video_info = author_video_info[["user.id", "avg_videos_per_week"]]

author_video_info.shape

Merge dữ liệu về user với dữ liệu về video

user_df = pd.merge(left=user_df, right=author_video_info, on="user.id", how="left")

### Tính một số giá trị trung bình khác trên mỗi video của user

Ta sẽ tính thêm một số giá trị khác như:

- Số lượt lưu video trung bình trên mỗi video (video - `statsV2.collectCount`)
- Số lượt bình luận trung bình trên mỗi video (video - `statsV2.commentCount`)
- Số lượt thích trung bình trên mỗi video (video - `statsV2.diggCount`)
- Số lượt xem trung bình trên mỗi video (video - `statsV2.playCount`)
- Số lượt chia sẻ trung bình trên mỗi video (video - `statsV2.shareCount`)
- Thời lượng trung bình của video (video - `video.duration`)
- Số lượng hashtag trung bình trên mỗi video (video - `hashtag_count`)


# Define interesting columns
interesting_columns = [
    "statsV2.collectCount", "statsV2.commentCount", "statsV2.diggCount",
    "statsV2.playCount", "statsV2.shareCount", "video.duration",
    "hashtag_count"
]

# Group DataFrame by `author.id` then calculate the mean
# of some interesting columns
avg_user_stats = video_df.groupby("author.id")[interesting_columns]\
                         .mean().reset_index()

# Rename the columns to match the user_df
avg_user_stats = avg_user_stats.rename(
    columns={
        "author.id": "user.id",
        "statsV2.collectCount": "avg_collects_per_video",
        "statsV2.commentCount": "avg_comments_per_video",
        "statsV2.diggCount": "avg_diggs_per_video",
        "statsV2.playCount": "avg_plays_per_video",
        "statsV2.shareCount": "avg_shares_per_video",
        "video.duration": "avg_video_duration",
        "hashtag_count": "avg_hashtags_per_video"
    }
)

# Merge the two DataFrames to get the average stats along with the author video info
user_df = pd.merge(
    left=user_df, right=avg_user_stats, on="user.id", how="left")

user_df.shape

## Lữu dữ liệu thành file Parquet

# Save the cleaned user_df to a new parquet file
user_df.to_parquet("./data/processed/cleaned_user_info.parquet", index=False)

pd.read_parquet(
    "./data/processed/cleaned_user_info.parquet").info()

- `stats.diggCount`: Số lượng video hoặc bình luận mà 1 người dùng đã "like"

# ----------------------

# END

# ----------------------


import pandas as pd

df = pd.read_parquet(
    "data/processed/cleaned_user_info.parquet")
df.info()

df['avg_videos_per_week'].mean()

stats_cols = df.columns[df.columns.str.contains("stats")]
df[stats_cols].describe()

df["stats.diggCount"].value_counts()
