==> notebooks/02_DEP_03-Filter_users.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Đọc dữ liệu từ file 

df = pd.read_parquet("./preprocessed_videos.parquet")
df.info()

# Đọc danh sách filter cho author

filtered_authors = None
with open("./data/filters/author_unique_ids.txt", "r") as f:
    filtered_authors = f.read().splitlines()

print(f"Number of filtered authors: {len(filtered_authors)}")

# Print number of videos per author
for author_id in filtered_authors:
    url = f"https://www.tiktok.com/@{author_id}"

    # Print author_id, number of videos, and URL
    print(f"Author: {author_id}")
    print(
        f"Number of videos: {df[df['author.uniqueId'] == author_id].shape[0]}")
    print(f"URL: {url}")
    print()

# Lọc dữ liệu theo danh sách filter


n_videos = 298 + 255 + 104 + 258 + 169 + 251
print(f"Total number of videos: {n_videos}")

# Filter row with filtered authors
# then reset index to avoid missing index
filtered_authors_df = df[df["author.uniqueId"].isin(filtered_authors)]
filtered_authors_df = filtered_authors_df.reset_index(drop=True)
assert filtered_authors_df.shape[0] == n_videos
filtered_authors_df.info()

# Save data to parquet, without index
filtered_authors_df.to_parquet("./filtered_authors.parquet", index=False)

pd.read_parquet("./filtered_authors.parquet").info()
