==> notebooks/03_FE_02_01-Count_food_location_videos.ipynb <==
# Tìm ra các file JSON trong thư mục


from typing import List
import os


def list_file_types(directory: str, file_extension: str) -> List[str]:
    """ List all files with a specific extension in a directory.

    Args:
        directory (str): Directory path.
        file_extension (str): File extension.

    Returns:
        List[str]: List of file paths.
    """

    file_list: List[str] = []
    for root, dirs, files in os.walk(directory):
        for file in files:
            if file.endswith(file_extension):
                file_list.append(os.path.join(root, file))
    return file_list

json_files = list_file_types("../data/foods-and-locations", ".json")
print(f"Number of JSON files found: {len(json_files)}")
# print(json_files[:5])

from tqdm import tqdm

# Set of filenames
file_names = set()

# Get a list of JSON files
for json_file in tqdm(json_files):
    # Extract base name of the file
    # remove file extension
    file_name = os.path.basename(json_file).replace(".json", "")
    # print(f"Processing file: {file_name}")

    # Append to the set
    file_names.add(file_name)

assert len(file_names) == len(json_files), "Duplicate file names found!"

# Write the set of filenames to a text file
with open("preprocessed_video_ids.txt", "w") as f:
    for file_name in file_names:
        f.write(f"{file_name}\n")

# Giảm kích thước file Parquet

# import pandas as pd
# df = pd.read_parquet("top_20_percent_weekly_videos_transcripts.parquet")
# df[["video.id", "desc", "transcript"]].to_parquet(
#     "small_weekly_videos_transcripts.parquet", index=False)
